+++ scripts/local/create-dao.ts
@@ -2,6 +2,7 @@ import {ethers} from 'hardhat'
 import {BondMediator} from '../../typechain-types'
 import {log} from '../../config/logging'
 import {ContractReceipt, Event} from 'ethers'
+import {addressEnvironmentVariable} from '../utils/environment-variable'
 
 async function createDao(mediatorAddress: string, treasuryAddress: string) {
     const factory = await ethers.getContractFactory('BondMediator')
@@ -23,10 +24,8 @@ async function createDao(mediatorAddress: string, treasuryAddress: string) {
 }
 
 async function main(): Promise<void> {
-    const mediator = parseEnvironmentVariable('BOND_MEDIATOR_CONTRACT')
-    const treasury = parseEnvironmentVariable('TREASURY_ADDRESS')
-
-    // TODO validate input - addresses
+    const mediator = addressEnvironmentVariable('BOND_MEDIATOR_CONTRACT')
+    const treasury = addressEnvironmentVariable('TREASURY_ADDRESS')
 
     return createDao(mediator, treasury)
 }
@@ -34,19 +33,6 @@ async function main(): Promise<void> {
 function receiptEvents(receipt: ContractReceipt): Event[] {
     const availableEvents = receipt.events
     return availableEvents ? availableEvents : []
-}
-
-function parseEnvironmentVariable(name: string): string {
-    const envVar = process.env[name]
-
-    log.info(envVar)
-
-    // eslint-disable-next-line no-undefined
-    if (envVar === undefined) {
-        throw Error(`Missing environment variable: ${name}`)
-    }
-
-    return envVar
 }
 
 main()
